Decoupled the code a little - added event handlers instead of passing ScreenManager as an argument, moved some code from ScreenManager to RightSideManager and LeftSideManager

+++ src/main/java/org/example/mtgspotscrapper/viewmodel/eventHandling/handlers/DeleteListEventHandler.java
@@ -7,6 +7,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.sql.SQLException;
+import java.util.Objects;
 
 public class DeleteListEventHandler extends MyEventHandler<DeleteListEvent> {
     protected static Logger deleteListLogger = LoggerFactory.getLogger(DeleteListEventHandler.class);
@@ -18,10 +19,15 @@ public class DeleteListEventHandler extends MyEventHandler<DeleteListEvent> {
     @Override
     public void handle(DeleteListEvent deleteListEvent) {
         try {
+            if (Objects.equals(deleteListEvent.getListName(), "Collection")) {
+                deleteListLogger.info("Can't delete collection");
+                return;
+            }
+
             if(databaseService.deleteList(deleteListEvent.getListName()))
                 deleteListLogger.info("Deleted list: {}", deleteListEvent.getListName());
             else deleteListLogger.info("Could not delete list (list doesn't exist): {}", deleteListEvent.getListName());
-        } catch (SQLException e) {
+        } catch (Exception e) {
             deleteListLogger.error("Failed to delete list {}", deleteListEvent.getListName(), e);
             throw new RuntimeException(e);
         }
